Fix: Extract MockData from Components to avoid code Duplication flag
