+++ src/pages/JobProfile/components/JobProfileItem.jsx
@@ -1,114 +1,39 @@
 import { useTheme } from "@emotion/react";
-import {Grid, ListItem, ListItemText } from "@mui/material";
+import { Grid, ListItem, ListItemText } from "@mui/material";
+import mockedJobs from "../mockData";
 
-function JobProfileItem(){
-
-    const mockedJobs = [
-        {
-            category: "Web Development - 1", 
-            title: "Frontend Developer"
-        },
-        {
-            category: "Web Development - 2", 
-            title: "Full Stack Developer"
-        },
-        {
-            category: "Mobile App Development - 1", 
-            title: "iOS Developer"
-        },
-        {
-            category: "Mobile App Development - 2", 
-            title: "Android Developer"
-        },
-        {
-            category: "Data Science", 
-            title: "Data Scientist"
-        },
-        {
-            category: "Machine Learning", 
-            title: "Machine Learning Engineer"
-        },
-        {
-            category: "DevOps", 
-            title: "DevOps Engineer"
-        },
-        {
-            category: "Cybersecurity", 
-            title: "Cybersecurity Analyst"
-        },
-        {
-            category: "Game Development", 
-            title: "Game Developer"
-        },
-        {
-            category: "Cloud Computing", 
-            title: "Cloud Solutions Architect"
-        },
-        {
-            category: "Database Administration", 
-            title: "Database Administrator"
-        },
-        {
-            category: "UI/UX Design", 
-            title: "UI/UX Designer"
-        },
-        {
-            category: "Software Testing", 
-            title: "Quality Assurance Engineer"
-        },
-        {
-            category: "Blockchain Development", 
-            title: "Blockchain Developer"
-        },
-        {
-            category: "Systems Administration", 
-            title: "Systems Administrator"
-        },
-        {
-            category: "Artificial Intelligence", 
-            title: "AI Research Scientist"
-        },
-        {
-            category: "Augmented Reality", 
-            title: "AR Developer"
-        },
-        {
-            category: "Technical Writing", 
-            title: "Technical Writer"
-        }
-    ];
-
-    return (
-        <Grid container 
-        rowSpacing={6} 
-        columnSpacing={10} 
-        columns={16}
-        sx={{
-            display: "flex", 
-            justifyContent: "center",
-            marginTop: "50px",
-            marginBottom: "50px"
-        }}
-        >
-            {mockedJobs?.map((job) => (
-                <Grid item xs={6} key={job.title}>
-                    <ListItem
-                    sx={{
-                        height: "100px",
-                        background: useTheme().palette.primary.light,
-                    }}
-                    >
-                        <ListItemText
-                        primary={job.category}
-                        secondary={job.title}
-                        secondaryTypographyProps={{ align: "center" }}
-                        />
-                    </ListItem>
-                </Grid>
-            ))}
-                        
+function JobProfileItem() {
+  return (
+    <Grid
+      container
+      rowSpacing={6}
+      columnSpacing={10}
+      columns={16}
+      sx={{
+        display: "flex",
+        justifyContent: "center",
+        marginTop: "50px",
+        marginBottom: "50px",
+      }}
+    >
+      {mockedJobs?.map((job) => (
+        <Grid item xs={6} key={job.title}>
+          <ListItem
+            sx={{
+              height: "100px",
+              background: useTheme().palette.primary.light,
+            }}
+          >
+            <ListItemText
+              primary={job.category}
+              secondary={job.title}
+              secondaryTypographyProps={{ align: "center" }}
+            />
+          </ListItem>
         </Grid>
-        )
-    }
-            
+      ))}
+    </Grid>
+  );
+}
+
 export default JobProfileItem;
