+++ src/pages/JobProfile/mockData.js
@@ -1,0 +1,76 @@
+const mockedJobs = [
+  {
+    category: "Web Development - 1",
+    title: "Frontend Developer",
+  },
+  {
+    category: "Web Development - 2",
+    title: "Full Stack Developer",
+  },
+  {
+    category: "Mobile App Development - 1",
+    title: "iOS Developer",
+  },
+  {
+    category: "Mobile App Development - 2",
+    title: "Android Developer",
+  },
+  {
+    category: "Data Science",
+    title: "Data Scientist",
+  },
+  {
+    category: "Machine Learning",
+    title: "Machine Learning Engineer",
+  },
+  {
+    category: "DevOps",
+    title: "DevOps Engineer",
+  },
+  {
+    category: "Cybersecurity",
+    title: "Cybersecurity Analyst",
+  },
+  {
+    category: "Game Development",
+    title: "Game Developer",
+  },
+  {
+    category: "Cloud Computing",
+    title: "Cloud Solutions Architect",
+  },
+  {
+    category: "Database Administration",
+    title: "Database Administrator",
+  },
+  {
+    category: "UI/UX Design",
+    title: "UI/UX Designer",
+  },
+  {
+    category: "Software Testing",
+    title: "Quality Assurance Engineer",
+  },
+  {
+    category: "Blockchain Development",
+    title: "Blockchain Developer",
+  },
+  {
+    category: "Systems Administration",
+    title: "Systems Administrator",
+  },
+  {
+    category: "Artificial Intelligence",
+    title: "AI Research Scientist",
+  },
+  {
+    category: "Augmented Reality",
+    title: "AR Developer",
+  },
+  {
+    category: "Technical Writing",
+    title: "Technical Writer",
+  },
+];
+
+export default mockedJobs;
